sem pog de isUpdate!

+++ loan/src/ifpr/competicao/jogos/mb/JogosMB.java
@@ -81,8 +81,6 @@ public class JogosMB {
 
 	private JogosTime jogosTime;
 
-	private boolean isUpdate;
-
 	@ManagedProperty(value = "#{relatorioFinal}")
 	public RelatorioFinal relatorioFinal;
 
@@ -91,12 +89,10 @@ public class JogosMB {
 		timesSelecionados = new ArrayList<>();
 		campusSelecionados = new ArrayList<>();
 		modalidadesSelecionadas = new ArrayList<>();
-		isUpdate = true;
 	}
 
 	public void criar() {
 		jogos = new Jogos();
-		isUpdate = false;
 	}
 
 	@PostConstruct
@@ -107,10 +103,7 @@ public class JogosMB {
 	}
 
 	public void cancelar() {
-		if (isUpdate == false) {
-			removerRegistrosRelacionamentos();
-			isUpdate = true;
-		}
+		removerRegistrosRelacionamentos();
 		jogos = null;
 
 	}
@@ -118,13 +111,19 @@ public class JogosMB {
 	private void removerRegistrosRelacionamentos() {
 		try {
 			for (JogosTime jogosTime : jogos.getJogosTimes()) {
-				jogosTimeDao.remover(jogosTime);
+				if (jogosTime.getJogos() == null) {
+					jogosTimeDao.remover(jogosTime);
+				}
 			}
 			for (JogosCampus jogosCampus : jogos.getJogosCampus()) {
-				jogosCampusDao.remover(jogosCampus);
+				if (jogosCampus.getJogos() == null) {
+					jogosCampusDao.remover(jogosCampus);
+				}
 			}
 			for (JogosModalidade jogosModalidade : jogos.getJogosModalidades()) {
-				jogosModalidadeDao.remover(jogosModalidade);
+				if (jogosModalidade.getJogos() == null) {
+					jogosModalidadeDao.remover(jogosModalidade);
+				}
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -136,36 +135,65 @@ public class JogosMB {
 
 		JogosTime jogosTime;
 		for (Time time : timesSelecionados) {
-			jogosTime = new JogosTime();
-			jogosTime.setTime(time);
-			jogosTimeDao.salvar(jogosTime);
-			jogos.getJogosTimes().add(jogosTime);
+			if (jogosTimeExiste(time)) {
+				jogosTime = new JogosTime();
+				jogosTime.setTime(time);
+				jogosTimeDao.salvar(jogosTime);
+				jogos.getJogosTimes().add(jogosTime);
+			}
 		}
-
 	}
 
 	public void adicionarCampus() {
-
 		JogosCampus jogosCampus;
 		for (Campus campus : campusSelecionados) {
-			jogosCampus = new JogosCampus();
-			jogosCampus.setCampus(campus);
-			jogosCampusDao.salvar(jogosCampus);
-			jogos.getJogosCampus().add(jogosCampus);
+			if (jogosCampusExiste(campus)) {
+				jogosCampus = new JogosCampus();
+				jogosCampus.setCampus(campus);
+				jogosCampusDao.salvar(jogosCampus);
+				jogos.getJogosCampus().add(jogosCampus);
+			}
+		}
+	}
+
+	public void adicionarModalidade() {
+		JogosModalidade jogosModalidade;
+		for (Modalidade modalidade : modalidadesSelecionadas) {
+			if (jogosModalidadeExiste(modalidade)) {
+				jogosModalidade = new JogosModalidade();
+				jogosModalidade.setModalidade(modalidade);
+				jogosModalidadeDao.salvar(jogosModalidade);
+				jogos.getJogosModalidades().add(jogosModalidade);
+			}
 		}
 
 	}
 
-	public void adicionarModalidade() {
-
-		JogosModalidade jogosModalidade;
-		for (Modalidade modalidade : modalidadesSelecionadas) {
-			jogosModalidade = new JogosModalidade();
-			jogosModalidade.setModalidade(modalidade);
-			jogosModalidadeDao.salvar(jogosModalidade);
-			jogos.getJogosModalidades().add(jogosModalidade);
+	private boolean jogosTimeExiste(Time time) {
+		for (JogosTime jogosTime : jogos.getJogosTimes()) {
+			if (jogosTime.getTime().equals(time)) {
+				return false;
+			}
 		}
+		return true;
+	}
 
+	private boolean jogosCampusExiste(Campus campus) {
+		for (JogosCampus jogosCampus : jogos.getJogosCampus()) {
+			if (jogosCampus.getCampus().equals(campus)) {
+				return false;
+			}
+		}
+		return true;
+	}
+
+	private boolean jogosModalidadeExiste(Modalidade modalidade) {
+		for (JogosModalidade jogosModalidade : jogos.getJogosModalidades()) {
+			if (jogosModalidade.getModalidade().equals(modalidade)) {
+				return false;
+			}
+		}
+		return true;
 	}
 
 	public void salvarJogos() {
